Updated for newer version of R and miniforge!
Also, added a couple of options that can help with debugging the heredoc
RUN commands. `set -e` errors out if the command failed and `set -x`
prints the expanded command that is executed which is very useful for
debugging issues!

This update also required setting an envvar that is used by the
reticulate library.