Use default values for ARGs
It was an oversight on my part not to have default values for the uid/gid related ARG variables. On my computer i build the image using make where i explicitly set the uid/gid arg vars at the cmdline.  However, that is not the case when the Dockerfile is used standalone, especially on a windows computer which does not have the concept of uid/gid.